refact: resolve warn (react-refresh/only-export-components)
Arquivos que exportam componentes não podem exportar outras funções e contantes.
