+++ src/pages/code.page.tsx
@@ -1,6 +1,6 @@
 import { routes } from "../links";
 import { Link } from "../router/link";
-import { useLinks } from "../router/router";
+import { useLinks } from "../router/hook.ts";
 
 export default function CodePage() {
   const links = useLinks<typeof routes.routes>();
